EAGLE-474 some improvements of app framework: appId naming, jarFile field, InstallOperation etc. improvements of app framework: appId naming, jarFile field, InstallOperation
https://issues.apache.org/jira/browse/EAGLE-474

Author: @yonzhang <yonzhang2012@apache.org>

Closes: #362
