+++ eagle-core/eagle-app/eagle-app-base/src/main/java/org/apache/eagle/app/spi/AbstractApplicationProvider.java
@@ -81,7 +81,6 @@ public abstract class AbstractApplicationProvider<T extends Application> impleme
 
     @Override
     public void prepare(ApplicationProviderConfig providerConfig, Config envConfig) {
-        this.applicationDesc.setJarPath(providerConfig.getJarPath());
     }
 
     protected void setVersion(String version) {
