+++ eagle-security/eagle-security-hbase-auditlog/src/main/resources/scripts
@@ -13,10 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-./kafka-topics.sh --topic sandbox_hbase_audit_log --zookeeper localhost:2181 --create --replication-factor 1 --partitions 2
+./kafka-topics.sh --topic hbase_audit_log --zookeeper localhost:2181 --create --replication-factor 1 --partitions 2
 
-./kafka-topics.sh --topic sandbox_hbase_audit_log_parsed --zookeeper localhost:2181 --create --replication-factor 1 --partitions 2
+./kafka-topics.sh --topic hbase_audit_log --zookeeper localhost:2181 --create --replication-factor 1 --partitions 2
 
-./kafka-console-producer.sh --topic sandbox_hbase_audit_log --broker-list sandbox.hortonworks.com:6667
+./kafka-console-producer.sh --topic hbase_audit_log --broker-list sandbox.hortonworks.com:6667
 
-./kafka-console-consumer.sh --topic sandbox_hbase_audit_log_parsed --zookeeper sandbox.hortonworks.com:2181 --from-beginning
+./kafka-console-consumer.sh --topic hbase_audit_log_parsed --zookeeper sandbox.hortonworks.com:2181 --from-beginning
