+++ eagle-server/src/main/resources/application.conf
@@ -53,7 +53,7 @@
 			"type": "org.apache.eagle.app.sink.KafkaStreamSink"
 		},
 		"storm": {
-			"nimbusHost": "localhost"
+			"nimbusHost": "server.eagle.apache.org"
 			"nimbusThriftPort": 6627
 		},
 		"provider" : {
