+++ eagle-topology-assembly/src/assembly/eagle-topology-assembly.xml
@@ -41,6 +41,7 @@
             </unpackOptions>
             <excludes>
                 <exclude>org.slf4j:slf4j-api</exclude>
+                <exclude>org.slf4j:slf4j-log4j12</exclude>
                 <exclude>io.dropwizard:**</exclude>
                 <exclude>com.sun.jersey:**</exclude>
                 <exclude>com.sun.jersey.contribs:**</exclude>
